refactor: extract runtime core contracts
