+++ packages/runtime-core/src/runtime-episode-contracts.ts
@@ -12,7 +12,7 @@ import type {
   RuntimeEpisodeTraceValidationIssue,
   RuntimeEpisodeTraceValidationResult,
   Snapshot,
-} from "./index.js"
+} from "./runtime-contracts.js"
 
 export const RUNTIME_EPISODE_TRACE_SCHEMA = "wp-codebox/runtime-episode-trace/v1" as const
 export const RUNTIME_EPISODE_ACTION_SCHEMA = "wp-codebox/runtime-episode-action/v1" as const
